pyproject.toml
[tool.poetry]
name = "gitjob-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
SQLAlchemy = "^2.0.29"
python-dotenv = "^1.0.1"
uvicorn = "^0.29.0"
psycopg2-binary = "^2.9.9"
httpx = "^0.27.0"
testing-postgresql = "^1.3.0"
email-validator = "^2.1.1"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
pytz = "^2024.1"
dirty-equals = "^0.7.1.post0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
isort = "^5.13.2"

[tool.poetry.group.pytest]
optional = true

[tool.poetry.group.pytest.dependencies]
pytest = "^8.1.1"
pytest-md = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
